Fix 'Remove Folder' and 'Add Folder' commands
Fixes #5070

Since when building the `navigator` context menu the commands
are not yet registered due to waiting for the preferences to be
ready, a placeholder command is registered, and then un-registered
and re-registered with the proper implementations.

This means that building menus will not fail as the placeholder command
is properly registered, and then when ready, the command is unregistered, and
re-registered with the correct implementation.

Additional Fixes:
- remove unnecessary `this.updateAddRemoveFolderActions(registry)` call.
- fix typo from `toDispoaseAddRemoveFolderActions` to `toDisposeAddRemoveFolderActions`.

Signed-off-by: Vincent Fugnitto <vincent.fugnitto@ericsson.com>